refactor(research): use app-owned sha256 helper
