+++ src/research/source-epub-comparator.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { sha256HexSync } from '../struct/sha256'
+import { sha256HexSync } from './sha256-sync'
 import {
   DETERMINISTIC_CHECK_IDS,
   encodeCanonicalObservationPayload,
